Сделал 5 задание по модулям. Не работает вывод в консоль
